# docs/api/openapi.v1.json
{
  "components": {
    "schemas": {
      "AddCommentRequest": {
        "properties": {
          "author": {
            "type": [
              "null",
              "string"
            ]
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content",
          "author"
        ],
        "type": "object"
      },
...
      "AssignTicketRequest": {
        "properties": {
          "agentId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "agentId"
        ],
        "type": "object"
      },
...
      "AuditEventResponse": {
        "properties": {
          "actor": {
            "type": [
              "null",
              "string"
            ]
          },
          "eventType": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/AuditEventPayload"
          }
        },
        "required": [
          "id",
          "eventType",
          "actor",
          "occurredAt",
          "payload"
        ],
        "type": "object"
      },
...
      "CommentResponse": {
        "properties": {
          "author": {
            "type": [
              "null",
              "string"
            ]
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "author",
          "createdAt"
        ],
        "type": "object"
      },
...
      "CreateTicketRequest": {
        "properties": {
          "description": {
            "type": "string"
          },
          "dueAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "priority",
          "dueAt"
        ],
        "type": "object"
      },
      "CreateTicketResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
...
      "PagedResponseOfTicketListItemResponse": {
        "properties": {
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TicketListItemResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "pageSize": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "totalCount": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "totalPages": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object"
      },
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
      "SortDirection": {
        "enum": [
          "Asc",
          "Desc",
          null
        ]
      },
...
      "TicketPriority": {
        "enum": [
          "Low",
          "Medium",
          "High",
          "Critical"
        ],
        "type": "string"
      },
      "TicketResponse": {
        "properties": {
          "allowedTransitions": {
            "items": {
              "$ref": "#/components/schemas/TicketStatus"
            },
            "type": "array"
          },
          "assignee": {
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "$ref": "#/components/schemas/TicketCategory"
          },
          "conversation": {
            "items": {
              "$ref": "#/components/schemas/ConversationItemResponse"
            },
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "displayRef": {
            "type": "string"
          },
          "dueAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isOverdue": {
            "type": "boolean"
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "statusGuidance": {
            "type": "string"
          },
          "suggestedNextSteps": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "priority",
          "category",
          "status",
          "createdAt",
          "dueAt",
          "assignee",
          "conversation",
          "allowedTransitions",
          "isOverdue",
          "displayRef",
          "statusGuidance",
          "suggestedNextSteps"
        ],
        "type": "object"
      },
      "TicketSortField": {
        "enum": [
          "CreatedAt",
          "DueAt",
          "Priority",
          "Status",
          "Title",
          null
        ]
      },
      "TicketStatus": {
        "enum": [
          "New",
          "Triaged",
          "InProgress",
          "Waiting",
          "Resolved",
          "Closed"
        ],
        "type": "string"
      },
      "UpdateTicketRequest": {
        "properties": {
          "category": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TicketCategory"
              }
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "dueAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "priority": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ServiceDeskLite.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/tickets": {
      "get": {
        "description": "Returns a paged list of tickets with deterministic sorting (CreatedAt + Id).",
        "operationId": "Tickets_Search",
        "parameters": [
          {
            "in": "query",
            "name": "Page",
            "schema": {
              "default": 1,
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "PageSize",
            "schema": {
              "default": 25,
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "SortField",
            "schema": {
              "$ref": "#/components/schemas/TicketSortField"
            }
          },
          {
            "in": "query",
            "name": "SortDirection",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          },
          {
            "in": "query",
            "name": "Q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "Statuses",
            "schema": {
              "items": {
                "$ref": "#/components/schemas/TicketStatus"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "Priorities",
            "schema": {
              "items": {
                "$ref": "#/components/schemas/TicketPriority"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "Assignee",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfTicketListItemResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Search tickets",
        "tags": [
          "Tickets"
        ]
      },
      "post": {
        "operationId": "Tickets_Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTicketResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create a new ticket",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/api/v1/tickets/{id}": {
      "get": {
        "operationId": "Tickets_GetById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get ticket by id",
        "tags": [
          "Tickets"
        ]
      },
      "patch": {
        "description": "Partial update of title, description, priority, or due date. Omitted fields stay unchanged.",
        "operationId": "Tickets_Update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update ticket details",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/api/v1/tickets/{id}/assign": {
      "post": {
        "description": "Set AssigneeName to null to unassign.",
        "operationId": "Tickets_Assign",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Assign or unassign a ticket",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/api/v1/tickets/{id}/audit-events": {
      "get": {
        "description": "Returns all audit events for a ticket in chronological order.",
        "operationId": "Tickets_GetAuditEvents",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get audit history for a ticket",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/api/v1/tickets/{id}/comments": {
      "post": {
        "operationId": "Tickets_AddComment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Add a comment to a ticket",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/api/v1/tickets/{id}/status": {
      "post": {
        "operationId": "Tickets_ChangeStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeTicketStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Change ticket status",
        "tags": [
          "Tickets"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5238/"
    }
  ],
...
}